+++ mx.truffle/mx_truffle.py
@@ -27,8 +27,8 @@
 # ----------------------------------------------------------------------------------------------------
 
 import os
+from os.path import exists
 import re
-import subprocess
 
 import mx
 
@@ -38,6 +38,7 @@ from mx_sigtest import sigtest
 from mx_jackpot import jackpot
 from mx_gate import Task
 from mx_javamodules import as_java_module, get_java_module_info
+from urlparse import urljoin
 import mx_gate
 import mx_unittest
 
@@ -46,12 +47,61 @@ _suite = mx.suite('truffle')
 def javadoc(args, vm=None):
     """build the Javadoc for all API packages"""
     mx.javadoc(['--unified'] + args)
-    index = os.sep.join([_suite.dir, 'javadoc', 'index.html'])
+    javadocDir = os.sep.join([_suite.dir, 'javadoc'])
+    index = os.sep.join([javadocDir, 'index.html'])
     indexContent = open(index, 'r').read()
     indexContent = indexContent.replace('src="allclasses-frame.html"', 'src="com/oracle/truffle/api/vm/package-frame.html"')
     indexContent = indexContent.replace('src="overview-summary.html"', 'src="com/oracle/truffle/tutorial/package-summary.html"')
     new_file = open(index, "w")
     new_file.write(indexContent)
+    checkLinks(javadocDir)
+
+def checkLinks(javadocDir):
+    href = re.compile('(?<=href=").*?(?=")')
+    filesToCheck = {}
+    for root, _, files in os.walk(javadocDir):
+        for f in files:
+            if f.endswith('.html'):
+                html = os.path.join(root, f)
+                content = open(html, 'r').read()
+                for url in href.findall(content):
+                    full = urljoin(html, url)
+                    sectionIndex = full.find('#')
+                    questionIndex = full.find('?')
+                    minIndex = sectionIndex
+                    if minIndex < 0:
+                        minIndex = len(full)
+                    if questionIndex >= 0 and questionIndex < minIndex:
+                        minIndex = questionIndex
+                    path = full[0:minIndex]
+
+                    sectionNames = filesToCheck.get(path, [])
+                    if sectionIndex >= 0:
+                        s = full[sectionIndex + 1:]
+                        sectionNames = sectionNames + [(html, s)]
+                    else:
+                        sectionNames = sectionNames + [(html, None)]
+
+                    filesToCheck[path] = sectionNames
+
+    err = False
+    for referencedfile, sections in filesToCheck.items():
+        if referencedfile.startswith('javascript:') or referencedfile.startswith('http:') or referencedfile.startswith('https:') or referencedfile.startswith('mailto:'):
+            continue
+        if not exists(referencedfile):
+            mx.warn('Referenced file ' + referencedfile + ' does not exist. Referenced from ' + sections[0][0])
+            err = True
+        else:
+            content = open(referencedfile, 'r').read()
+            for path, s in sections:
+                if not s == None:
+                    where = content.find('name="' + s + '"')
+                    if where == -1:
+                        mx.warn('There should be section ' + s + ' in ' + referencedfile + ". Referenced from " + path)
+                        err = True
+
+    if err:
+        mx.abort('There are wrong references in Javadoc')
 
 def build(args, vm=None):
     """build the Java sources"""
@@ -109,25 +159,6 @@ def repl(args):
     vmArgs, slArgs = mx.extract_VM_args(args, useDoubleDash=True)
     mx.run_java(vmArgs + _path_args() + ["com.oracle.truffle.tools.debug.shell.client.SimpleREPLClient"] + slArgs)
 
-def testrubytruffle(args):
-    """test RubyTruffle against the Truffle API"""
-
-    jruby_dir = 'jruby'
-    jruby_repo = 'https://github.com/jruby/jruby.git'
-    jruby_branch = 'truffle-head'
-    git = mx.GitConfig()
-    if os.path.exists('jruby'):
-        git.run(['git', 'reset', 'HEAD', '--hard'], nonZeroIsFatal=True, cwd=jruby_dir)
-        git.pull('jruby')
-    else:
-        git.clone(jruby_repo, jruby_dir)
-        git.run(['git', 'checkout', jruby_branch], nonZeroIsFatal=True, cwd=jruby_dir)
-    dev_version = _suite.release_version(snapshotSuffix='SNAPSHOT')
-    mx.build([])
-    mx.maven_install([])
-    subprocess.check_call(['./mvnw', '-q', '-Dtruffle.version=' + dev_version], cwd=jruby_dir)
-    subprocess.check_call(['bin/jruby', 'tool/jt.rb', 'test', 'fast'], cwd=jruby_dir)
-
 def _truffle_gate_runner(args, tasks):
     jdk = mx.get_jdk(tag=mx.DEFAULT_JDK_TAG)
     with Task('Jackpot check', tasks) as t:
@@ -146,7 +177,6 @@ mx.update_commands(_suite, {
     'javadoc' : [javadoc, '[SL args|@VM options]'],
     'sl' : [sl, '[SL args|@VM options]'],
     'repl' : [repl, '[REPL Debugger args|@VM options]'],
-    'testrubytruffle' : [testrubytruffle, ''],
     'testgraal' : [testgraal, ''],
 })