+++ truffle/com.oracle.truffle.api.vm/src/com/oracle/truffle/api/vm/PolyglotEngine.java
@@ -48,25 +48,19 @@ import java.util.logging.Logger;
 
 import com.oracle.truffle.api.CallTarget;
 import com.oracle.truffle.api.CompilerAsserts;
-import com.oracle.truffle.api.CompilerDirectives;
-import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
-import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.TruffleLanguage;
 import com.oracle.truffle.api.TruffleLanguage.Env;
-import com.oracle.truffle.api.frame.MaterializedFrame;
-import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.impl.Accessor;
 import com.oracle.truffle.api.impl.FindContextNode;
 import com.oracle.truffle.api.instrumentation.TruffleInstrument;
-import com.oracle.truffle.api.interop.ForeignAccess;
 import com.oracle.truffle.api.interop.Message;
 import com.oracle.truffle.api.interop.TruffleObject;
 import com.oracle.truffle.api.interop.java.JavaInterop;
-import com.oracle.truffle.api.nodes.DirectCallNode;
-import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.nodes.RootNode;
 import com.oracle.truffle.api.source.Source;
+import com.oracle.truffle.api.vm.ComputeInExecutor.Info;
 import com.oracle.truffle.api.vm.PolyglotEngine.Value;
+import com.oracle.truffle.api.vm.PolyglotRootNode.EvalRootNode;
 
 /**
  * A multi-language execution environment for Truffle-implemented {@linkplain Language languages}
@@ -213,6 +207,8 @@ public class PolyglotEngine {
     private final List<Object[]> config;
     private final Object[] debugger = {null};
     private final ContextStore context;
+    private final PolyglotCache cachedTargets;
+
     private volatile boolean disposed;
 
     static final boolean JDK8OrEarlier = System.getProperty("java.specification.version").compareTo("1.9") < 0;
@@ -244,6 +240,7 @@ public class PolyglotEngine {
         this.instruments = null;
         this.config = null;
         this.context = null;
+        this.cachedTargets = null;
     }
 
     /**
@@ -274,6 +271,11 @@ public class PolyglotEngine {
         this.langs = map;
         this.instruments = createAndAutostartDescriptors(InstrumentCache.load());
         this.context = ExecutionImpl.createStore(this);
+        this.cachedTargets = new PolyglotCache(this);
+    }
+
+    Info executor() {
+        return executor;
     }
 
     private Map<String, Instrument> createAndAutostartDescriptors(List<InstrumentCache> instrumentCaches) {
@@ -461,7 +463,7 @@ public class PolyglotEngine {
          * other than the calling thread.
          * <p>
          * By default engines execute both {@link PolyglotEngine#eval(Source)} and
-         * {@link Value#invoke(Object, Object[])} synchronously in the calling thread.
+         * {@link Value#execute(java.lang.Object...)} synchronously in the calling thread.
          * <p>
          * A custom {@link Executor} is expected to perform every execution it is given (via
          * {@link Executor#execute(Runnable)}) in order of arrival. An arbitrary thread may be used,
@@ -631,7 +633,7 @@ public class PolyglotEngine {
     private Object evalImpl(TruffleLanguage[] langTarget, Language l, Source source) {
         CallTarget target = l.cache.get(source);
         if (target == null) {
-            target = Truffle.getRuntime().createCallTarget(new PolyglotEvalRootNode(this, l, source));
+            target = PolyglotRootNode.createEval(this, l, source);
             l.cache.put(source, target);
         }
         Object value = target.call((Object) langTarget);
@@ -658,51 +660,8 @@ public class PolyglotEngine {
         return debugger;
     }
 
-    @SuppressWarnings({"try"})
-    final Object invokeForeign(final Node foreignNode, VirtualFrame frame, final TruffleObject receiver) {
-        assertNoTruffle();
-        Object res;
-        CompilerAsserts.neverPartOfCompilation();
-        if (executor == null) {
-            ContextStore prev = ExecutionImpl.executionStarted(context);
-            try {
-                Access.DEBUG.executionStarted(PolyglotEngine.this);
-                final Object[] args = ForeignAccess.getArguments(frame).toArray();
-                SymbolInvokerImpl.unwrapArgs(this, args);
-                res = ForeignAccess.execute(foreignNode, frame, receiver, args);
-            } finally {
-                ExecutionImpl.executionEnded(prev);
-            }
-        } else {
-            res = invokeForeignOnExecutor(foreignNode, frame, receiver);
-        }
-        return EngineTruffleObject.wrap(this, res);
-    }
-
     static void assertNoTruffle() {
         CompilerAsserts.neverPartOfCompilation("Methods of PolyglotEngine must not be compiled by Truffle. Use Truffle interoperability or a @TruffleBoundary instead.");
-    }
-
-    @TruffleBoundary
-    private Object invokeForeignOnExecutor(final Node foreignNode, VirtualFrame frame, final TruffleObject receiver) {
-        final MaterializedFrame materialized = frame.materialize();
-        ComputeInExecutor<Object> compute = new ComputeInExecutor<Object>(executor) {
-            @SuppressWarnings("try")
-            @Override
-            protected Object compute() {
-                ContextStore prev = ExecutionImpl.executionStarted(context);
-                try {
-                    Access.DEBUG.executionStarted(PolyglotEngine.this);
-                    final Object[] args = ForeignAccess.getArguments(materialized).toArray();
-                    RootNode node = SymbolInvokerImpl.createTemporaryRoot(TruffleLanguage.class, foreignNode, receiver);
-                    final CallTarget target = Truffle.getRuntime().createCallTarget(node);
-                    return target.call(args);
-                } finally {
-                    ExecutionImpl.executionEnded(prev);
-                }
-            }
-        };
-        return compute.get();
     }
 
     /**
@@ -923,77 +882,12 @@ public class PolyglotEngine {
         }
     }
 
-    abstract static class PolyglotRootNode extends RootNode {
-
-        protected final PolyglotEngine engine;
-
-        PolyglotRootNode(Class<? extends TruffleLanguage> language, PolyglotEngine engine) {
-            super(language, null, null);
-            this.engine = engine;
-        }
-
-        @Override
-        public final Object execute(VirtualFrame frame) {
-            ContextStore prev = ExecutionImpl.executionStarted(engine.context());
-            Access.DEBUG.executionStarted(engine);
-            try {
-                return executeImpl(frame);
-            } finally {
-                ExecutionImpl.executionEnded(prev);
-            }
-        }
-
-        protected abstract Object executeImpl(VirtualFrame frame);
-    }
-
-    private static class PolyglotEvalRootNode extends PolyglotRootNode {
-
-        private static final Object[] EMPTY_ARRAY = new Object[0];
-
-        @Child private DirectCallNode call;
-        private TruffleLanguage<?> fillLanguage;
-        private final Language language;
-        private final Source source;
-        private final PolyglotEngine engine;
-
-        PolyglotEvalRootNode(PolyglotEngine engine, Language language, Source source) {
-            super(TruffleLanguage.class, engine);
-            this.engine = engine;
-            this.source = source;
-            this.language = language;
-        }
-
-        public PolyglotEngine getEngine() {
-            return engine;
-        }
-
-        @Override
-        protected Object executeImpl(VirtualFrame frame) {
-            TruffleLanguage[] fillLang = (TruffleLanguage[]) frame.getArguments()[0];
-            if (call == null) {
-                CompilerDirectives.transferToInterpreterAndInvalidate();
-                initialize();
-            }
-            fillLang[0] = fillLanguage;
-            return call.call(EMPTY_ARRAY);
-        }
-
-        private void initialize() {
-            TruffleLanguage<?> languageImpl = language.getImpl(true);
-            CallTarget target = Access.LANGS.parse(languageImpl, source, null);
-            if (target == null) {
-                throw new NullPointerException("Parsing has not produced a CallTarget for " + source);
-            }
-            fillLanguage = languageImpl;
-            call = insert(DirectCallNode.create(target));
-        }
-    }
-
     /**
      * A future value wrapper. A user level wrapper around values returned by evaluation of various
      * {@link PolyglotEngine} functions like {@link PolyglotEngine#findGlobalSymbol(String)} and
-     * {@link PolyglotEngine#eval(Source)} or a value returned by {@link #invoke(Object, Object...)
-     * a subsequent execution}. In case the {@link PolyglotEngine} has been initialized for
+     * {@link PolyglotEngine#eval(Source)} or a value returned by
+     * {@link #execute(java.lang.Object...) a subsequent execution}. In case the
+     * {@link PolyglotEngine} has been initialized for
      * {@link Builder#executor(java.util.concurrent.Executor) asynchronous execution}, the
      * {@link Value} represents a future - i.e., it is returned immediately, leaving the execution
      * running on behind.
@@ -1002,7 +896,9 @@ public class PolyglotEngine {
      */
     public abstract class Value {
         private final TruffleLanguage<?>[] language;
-        private CallTarget target;
+        private CallTarget executeTarget;
+        private CallTarget unwrapTarget;
+        private CallTarget asJavaObjectTarget;
 
         Value(TruffleLanguage<?>[] language) {
             this.language = language;
@@ -1011,6 +907,31 @@ public class PolyglotEngine {
         abstract boolean isDirect();
 
         abstract Object value();
+
+        @SuppressWarnings("unchecked")
+        private <T> T unwrapJava(Object value) {
+            if (unwrapTarget == null) {
+                unwrapTarget = cachedTargets.lookupAsJava(value.getClass());
+            }
+            return (T) unwrapTarget.call(Object.class, value);
+        }
+
+        @SuppressWarnings("unchecked")
+        private <T> T asJavaObject(Class<T> type, Object value) {
+            if (asJavaObjectTarget == null) {
+                asJavaObjectTarget = cachedTargets.lookupAsJava(value == null ? void.class : value.getClass());
+            }
+            return (T) asJavaObjectTarget.call(type, value);
+        }
+
+        @SuppressWarnings("try")
+        private Object executeDirect(Object[] args) {
+            Object value = value();
+            if (executeTarget == null) {
+                executeTarget = cachedTargets.lookupExecute(value.getClass());
+            }
+            return executeTarget.call(value, args);
+        }
 
         /**
          * Returns the object represented by this value, possibly null. The <em>raw</em> object can
@@ -1023,18 +944,11 @@ public class PolyglotEngine {
          * @since 0.9
          */
         public Object get() {
-            return get(true, true);
-        }
-
-        private Object get(boolean unwrapJava, boolean wrapEngine) {
-            assertNoTruffle();
             Object result = waitForSymbol();
             if (result instanceof TruffleObject) {
-                if (unwrapJava) {
-                    result = JavaInterop.asJavaObject(Object.class, (TruffleObject) result);
-                }
-                if (wrapEngine && result instanceof TruffleObject) {
-                    return EngineTruffleObject.wrap(PolyglotEngine.this, result);
+                result = unwrapJava((TruffleObject) result);
+                if (result instanceof TruffleObject) {
+                    result = EngineTruffleObject.wrap(PolyglotEngine.this, result);
                 }
             }
             return result;
@@ -1068,26 +982,25 @@ public class PolyglotEngine {
          * @since 0.9
          */
         public <T> T as(final Class<T> representation) {
-            assertNoTruffle();
-            final Object obj = get(true, false);
-            if (obj instanceof EngineTruffleObject) {
-                EngineTruffleObject eto = (EngineTruffleObject) obj;
-                if (representation.isInstance(eto.getDelegate())) {
-                    return representation.cast(eto.getDelegate());
-                }
+            Object original = waitForSymbol();
+            Object unwrapped = original;
+
+            if (original instanceof TruffleObject) {
+                unwrapped = unwrapJava(original);
             }
             if (representation == String.class) {
                 final Class<? extends TruffleLanguage> clazz = language[0].getClass();
-                Object unwrapped = obj;
-                while (unwrapped instanceof EngineTruffleObject) {
-                    unwrapped = ((EngineTruffleObject) obj).getDelegate();
-                }
-                return representation.cast(Access.LANGS.toStringIfVisible(language[0], findEnv(clazz), unwrapped, null));
+                Object unwrappedConvered = unwrapped instanceof ConvertedObject ? ((ConvertedObject) unwrapped).getOriginal() : unwrapped;
+                return representation.cast(Access.LANGS.toStringIfVisible(language[0], findEnv(clazz), unwrappedConvered, null));
             }
-            if (representation.isInstance(obj)) {
-                return representation.cast(obj);
+            if (ConvertedObject.isInstance(representation, unwrapped)) {
+                return ConvertedObject.cast(representation, unwrapped);
             }
-            return JavaInterop.asJavaObject(representation, (TruffleObject) get(false, true));
+
+            if (original instanceof TruffleObject) {
+                original = EngineTruffleObject.wrap(PolyglotEngine.this, original);
+            }
+            return representation.cast(asJavaObject(representation, original));
         }
 
         /**
@@ -1194,18 +1107,13 @@ public class PolyglotEngine {
             return new ExecutorValue(language, invokeCompute);
         }
 
-        @SuppressWarnings("try")
-        private Object executeDirect(Object[] args) {
-            if (target == null) {
-                target = SymbolInvokerImpl.createExecuteSymbol(language[0], PolyglotEngine.this, value());
-            }
-            return target.call(args);
-        }
-
         private Object waitForSymbol() {
             assertNoTruffle();
             assert checkThread();
-            return value();
+            Object value = value();
+            assert value != null;
+            assert !(value instanceof EngineTruffleObject);
+            return value;
         }
     }
 
@@ -1215,6 +1123,7 @@ public class PolyglotEngine {
         DirectValue(TruffleLanguage<?>[] language, Object value) {
             super(language);
             this.value = value;
+            assert value != null;
         }
 
         @Override
@@ -1618,8 +1527,8 @@ public class PolyglotEngine {
 
             @Override
             public boolean isEvalRoot(RootNode target) {
-                if (target instanceof PolyglotEvalRootNode) {
-                    if (((PolyglotEvalRootNode) target).getEngine() == ExecutionImpl.findVM()) {
+                if (target instanceof EvalRootNode) {
+                    if (((EvalRootNode) target).getEngine() == ExecutionImpl.findVM()) {
                         return true;
                     }
                 }