+++ truffle/com.oracle.truffle.api/src/com/oracle/truffle/api/nodes/NodeClass.java
@@ -26,7 +26,9 @@ package com.oracle.truffle.api.nodes;
 
 import java.security.AccessController;
 import java.security.PrivilegedAction;
+import java.util.ArrayList;
 import java.util.Iterator;
+import java.util.List;
 
 /**
  * Information about a {@link Node} class. A single instance of this class is allocated for every
@@ -65,35 +67,53 @@ public abstract class NodeClass {
     /** @since 0.8 or earlier */
     @SuppressWarnings("deprecation")
     @Deprecated
-    public abstract NodeFieldAccessor getNodeClassField();
+    public NodeFieldAccessor getNodeClassField() {
+        return null;
+    }
 
     /** @since 0.8 or earlier */
     @SuppressWarnings("deprecation")
     @Deprecated
-    public abstract NodeFieldAccessor[] getCloneableFields();
+    public NodeFieldAccessor[] getCloneableFields() {
+        return null;
+    }
 
     /** @since 0.8 or earlier */
     @SuppressWarnings("deprecation")
     @Deprecated
-    public abstract NodeFieldAccessor[] getFields();
+    public NodeFieldAccessor[] getFields() {
+        return null;
+    }
 
     /** @since 0.8 or earlier */
     @SuppressWarnings("deprecation")
     @Deprecated
-    public abstract NodeFieldAccessor getParentField();
+    public NodeFieldAccessor getParentField() {
+        return null;
+    }
 
     /** @since 0.8 or earlier */
     @SuppressWarnings("deprecation")
     @Deprecated
-    public abstract NodeFieldAccessor[] getChildFields();
+    public NodeFieldAccessor[] getChildFields() {
+        return null;
+    }
 
     /** @since 0.8 or earlier */
     @SuppressWarnings("deprecation")
     @Deprecated
-    public abstract NodeFieldAccessor[] getChildrenFields();
+    public NodeFieldAccessor[] getChildrenFields() {
+        return null;
+    }
 
     /** @since 0.8 or earlier */
-    public abstract Iterator<Node> makeIterator(Node node);
+    public Iterator<Node> makeIterator(Node node) {
+        List<Object> arr = new ArrayList<>();
+        for (Object field : getNodeFields()) {
+            arr.add(field);
+        }
+        return new NodeIterator(this, node, arr.toArray());
+    }
 
     /**
      * The {@link Class} this <code>NodeClass</code> has been {@link #NodeClass(java.lang.Class)