+++ truffle/com.oracle.truffle.sl/src/com/oracle/truffle/sl/nodes/SLEvalRootNode.java
@@ -49,6 +49,7 @@ import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.source.SourceSection;
 import com.oracle.truffle.sl.SLLanguage;
 import com.oracle.truffle.sl.runtime.SLContext;
+import com.oracle.truffle.sl.runtime.SLNull;
 
 /**
  * In addition to {@link SLRootNode}, this class performs two additional tasks:
@@ -84,7 +85,7 @@ public final class SLEvalRootNode extends SLRootNode {
 
         if (getBodyNode() == null) {
             /* The source code did not have a "main" function, so nothing to execute. */
-            return null;
+            return SLNull.SINGLETON;
         }
 
         /* Conversion of arguments to types understood by SL. */