+++ truffle/com.oracle.truffle.tools.debug.shell/src/com/oracle/truffle/tools/debug/shell/server/REPLHandler.java
@@ -37,7 +37,6 @@ import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.source.Source;
 import com.oracle.truffle.api.source.SourceSection;
 import com.oracle.truffle.api.vm.PolyglotEngine.Language;
-import com.oracle.truffle.tools.debug.shell.server.InstrumentationUtils.ASTPrinter;
 import com.oracle.truffle.tools.debug.shell.server.InstrumentationUtils.LocationPrinter;
 import com.oracle.truffle.tools.debug.shell.server.REPLServer.BreakpointInfo;
 import com.oracle.truffle.tools.debug.shell.server.REPLServer.Context;
@@ -671,44 +670,6 @@ public abstract class REPLHandler {
         }
     };
 
-    public static final REPLHandler TRUFFLE_HANDLER = new REPLHandler(com.oracle.truffle.tools.debug.shell.REPLMessage.TRUFFLE) {
-
-        @Override
-        public com.oracle.truffle.tools.debug.shell.REPLMessage[] receive(com.oracle.truffle.tools.debug.shell.REPLMessage request, REPLServer replServer) {
-            final com.oracle.truffle.tools.debug.shell.REPLMessage reply = createReply();
-            final ASTPrinter astPrinter = replServer.getASTPrinter();
-            final String topic = request.get(com.oracle.truffle.tools.debug.shell.REPLMessage.TOPIC);
-            reply.put(com.oracle.truffle.tools.debug.shell.REPLMessage.TOPIC, topic);
-            Node node = replServer.getCurrentContext().getNodeAtHalt();
-            if (node == null) {
-                return finishReplyFailed(reply, "no current AST node");
-            }
-            final Integer depth = request.getIntValue(com.oracle.truffle.tools.debug.shell.REPLMessage.AST_DEPTH);
-            if (depth == null) {
-                return finishReplyFailed(reply, "missing AST depth");
-            }
-            try {
-                switch (topic) {
-                    case com.oracle.truffle.tools.debug.shell.REPLMessage.AST:
-                        while (node.getParent() != null) {
-                            node = node.getParent();
-                        }
-                        final String astText = astPrinter.displayAST(node, depth, replServer.getCurrentContext().getNodeAtHalt());
-                        return finishReplySucceeded(reply, astText);
-                    case com.oracle.truffle.tools.debug.shell.REPLMessage.SUBTREE:
-                    case com.oracle.truffle.tools.debug.shell.REPLMessage.SUB:
-                        final String subTreeText = astPrinter.displayAST(node, depth);
-                        return finishReplySucceeded(reply, subTreeText);
-                    default:
-                        return finishReplyFailed(reply, "Unknown \"" + com.oracle.truffle.tools.debug.shell.REPLMessage.TRUFFLE.toString() + "\" topic");
-                }
-
-            } catch (Exception ex) {
-                return finishReplyFailed(reply, ex);
-            }
-        }
-    };
-
     public static final REPLHandler UNSET_BREAK_CONDITION_HANDLER = new REPLHandler(com.oracle.truffle.tools.debug.shell.REPLMessage.UNSET_BREAK_CONDITION) {
 
         @Override
@@ -733,30 +694,4 @@ public abstract class REPLHandler {
         }
     };
 
-    public static final REPLHandler TRUFFLE_NODE_HANDLER = new REPLHandler(com.oracle.truffle.tools.debug.shell.REPLMessage.TRUFFLE_NODE) {
-
-        @Override
-        public com.oracle.truffle.tools.debug.shell.REPLMessage[] receive(com.oracle.truffle.tools.debug.shell.REPLMessage request, REPLServer replServer) {
-            final com.oracle.truffle.tools.debug.shell.REPLMessage reply = createReply();
-            final Node node = replServer.getCurrentContext().getNodeAtHalt();
-            if (node == null) {
-                return finishReplyFailed(reply, "no current AST node");
-            }
-
-            try {
-                final StringBuilder sb = new StringBuilder();
-                sb.append(replServer.getASTPrinter().displayNodeWithInstrumentation(node));
-                final SourceSection sourceSection = node.getSourceSection();
-                if (sourceSection != null) {
-                    final String code = sourceSection.getCode();
-                    sb.append(" \"");
-                    sb.append(code.substring(0, Math.min(code.length(), 15)));
-                    sb.append("...\"");
-                }
-                return finishReplySucceeded(reply, sb.toString());
-            } catch (Exception ex) {
-                return finishReplyFailed(reply, ex);
-            }
-        }
-    };
 }